+++ custom_components/git_ha_ppens/binary_sensor.py
@@ -53,7 +53,7 @@ class GitHaPpensDirtySensor(
             "name": "git-ha-ppens",
             "manufacturer": "git-ha-ppens",
             "model": "Git Version Control",
-            "sw_version": "1.0.1",
+            "sw_version": "1.1.0",
             "entry_type": "service",
             "configuration_url": "https://github.com/manuveli/git-ha-ppens",
         }